Changed timeSinceUpdate in interface to long

+++ src/annahack/telnetconnection/TelnetInterface.java
@@ -16,7 +16,7 @@ public interface TelnetInterface
 	public int getcursorx();
 	public int getcursory();
 	
-	public int timeSinceUpdate();
+	public long timeSinceUpdate();
 	
 	public boolean waiting() throws InterruptedException, IOException;
 }
